VM: moved EOF code to dedicated file, new constants for FORMAT, MAGIC and VERSION
