+++ packages/vm/src/evm/interpreter.ts
@@ -5,13 +5,8 @@ import { ERROR, VmError } from '../exceptions'
 import Memory from './memory'
 import Stack from './stack'
 import EEI from './eei'
-import {
-  Opcode,
-  handlers as opHandlers,
-  OpHandler,
-  AsyncOpHandler,
-  eof1CodeAnalysis,
-} from './opcodes'
+import { Opcode, handlers as opHandlers, OpHandler, AsyncOpHandler } from './opcodes'
+import * as eof from './opcodes/eof'
 import { dynamicGasHandlers } from './opcodes/gas'
 
 export interface InterpreterOpts {
@@ -93,21 +88,18 @@ export default class Interpreter {
   }
 
   async run(code: Buffer, opts: InterpreterOpts = {}): Promise<InterpreterResult> {
-    if (
-      !this._vm._common.isActivatedEIP(3540) ||
-      !code.slice(0, 1).equals(Buffer.from('ef', 'hex'))
-    ) {
+    if (!this._vm._common.isActivatedEIP(3540) || code[0] !== eof.FORMAT) {
       // EIP-3540 isn't active and first byte is not 0xEF - treat as legacy bytecode
       this._runState.code = code
     } else if (this._vm._common.isActivatedEIP(3540)) {
-      if (!code.slice(1, 2).equals(Buffer.from('00', 'hex'))) {
+      if (code[1] !== eof.MAGIC) {
         // Bytecode contains invalid EOF magic byte
         return {
           runState: this._runState,
           exceptionError: new VmError(ERROR.INVALID_BYTECODE_RESULT),
         }
       }
-      if (!code.slice(2, 3).equals(Buffer.from('01', 'hex'))) {
+      if (code[2] !== eof.VERSION) {
         // Bytecode contains invalid EOF version number
         return {
           runState: this._runState,
@@ -115,7 +107,7 @@ export default class Interpreter {
         }
       }
       // Code is EOF1 format
-      const codeSections = eof1CodeAnalysis(code)
+      const codeSections = eof.codeAnalysis(code)
       if (!codeSections) {
         // Code is invalid EOF1 format if `codeSections` is falsy
         return {
