+++ packages/vm/tests/api/EIPs/eip-3540-evm-object-format.spec.ts
@@ -3,7 +3,7 @@ import VM from '../../../src'
 import Common, { Chain, Hardfork } from '@ethereumjs/common'
 import { FeeMarketEIP1559Transaction } from '@ethereumjs/tx'
 import { Address, BN, privateToAddress } from 'ethereumjs-util'
-import { eof1CodeAnalysis } from '../../../src/evm/opcodes'
+import * as eof from '../../../src/evm/opcodes/eof'
 const pkey = Buffer.from('20'.repeat(32), 'hex')
 const GWEI = new BN('1000000000')
 const sender = new Address(privateToAddress(pkey))
@@ -16,14 +16,14 @@ tape('EIP 3540 tests', (t) => {
   })
 
   t.test('eof1CodeAnalysis() tests', async (st) => {
-    const eofHeader = Buffer.from([0xef, 0x00, 0x01])
+    const eofHeader = Buffer.from([eof.FORMAT, eof.MAGIC, eof.VERSION])
     st.ok(
-      eof1CodeAnalysis(Buffer.concat([eofHeader, Uint8Array.from([0x01, 0x00, 0x01, 0x00, 0x00])]))
+      eof.codeAnalysis(Buffer.concat([eofHeader, Uint8Array.from([0x01, 0x00, 0x01, 0x00, 0x00])]))
         ?.code! > 0,
       'valid code section'
     )
     st.ok(
-      eof1CodeAnalysis(
+      eof.codeAnalysis(
         Buffer.concat([
           eofHeader,
           Uint8Array.from([0x01, 0x00, 0x01, 0x02, 0x00, 0x01, 0x00, 0x00, 0xaa]),
@@ -32,7 +32,7 @@ tape('EIP 3540 tests', (t) => {
       'valid data section'
     )
     st.ok(
-      !eof1CodeAnalysis(
+      !eof.codeAnalysis(
         Buffer.concat([eofHeader, Uint8Array.from([0x01, 0x00, 0x01, 0x00, 0x00, 0x00])])
       ),
       'invalid container length'
